working text rendering

+++ src/font/Text.h
@@ -10,7 +10,7 @@ namespace font {
 	class Text {
 	public:
 		glm::vec2 position;
-		int size;
+		float size;
 		glm::vec3 color = glm::vec3(0.0f);
 		Font* font;
 		std::vector<glm::vec2> vertices, uvs;
@@ -20,7 +20,7 @@ namespace font {
 		std::string text;
 		
 	public:
-		Text(const std::string &text, const glm::vec2 &position, int size, Font *font);
+		Text(const std::string &text, const glm::vec2 &position, float size, Font *font);
 		void setText(const std::string &text);
 		std::vector<Character*> getCharacters();
